🔧 refactor: streamline authentication and search controllers
crawjud/addons/auth/__init__.py: Refactored AuthController to improve structure and added type hints.
crawjud/addons/auth/elaw.py: Updated ElawAuth to inherit from the refactored AuthController.
crawjud/addons/auth/esaj.py: Updated EsajAuth to inherit from the refactored AuthController.
crawjud/addons/auth/pje.py: Updated PjeAuth to inherit from the refactored AuthController.
crawjud/addons/auth/projudi.py: Updated ProjudiAuth to inherit from the refactored AuthController.
crawjud/addons/search/__init__.py: Refactored SearchController for better organization and type hints.
crawjud/addons/search/elaw.py: Updated ElawSearch to inherit from the refactored SearchController.
crawjud/addons/search/esaj.py: Updated EsajSearch to inherit from the refactored SearchController.
crawjud/addons/search/projudi.py: Updated ProjudiSearch to inherit from the refactored SearchController.
crawjud/core/master.py: Adjusted imports to reflect changes in AuthController and SearchController.
