+++ crawjud/addons/auth/__init__.py
@@ -1,27 +1,65 @@
-"""Módulo de autenticação CrawJUD."""
+"""Módulo controller de autenticação."""
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
-from crawjud.addons.auth.elaw import ElawAuth
-from crawjud.addons.auth.esaj import EsajAuth
-from crawjud.addons.auth.pje import PjeAuth
-from crawjud.addons.auth.projudi import ProjudiAuth
+from typing import TYPE_CHECKING, Self
 
 if TYPE_CHECKING:
-    from crawjud.addons.auth.controller import AuthController
+    from selenium.webdriver.remote.webdriver import WebDriver
+    from selenium.webdriver.support.wait import WebDriverWait
 
-auth_systems: type[AuthController] = {
-    "pje": PjeAuth,
-    "esaj": EsajAuth,
-    "elaw": ElawAuth,
-    "projudi": ProjudiAuth,
-}
+    from addons.printlogs import PrintMessage
+    from crawjud.types.elements import type_elements
 
 
-def authenticator(system: str) -> type[AuthController]:
-    """Retorna o objeto do autenticador."""
-    auth: type[AuthController] = auth_systems[system]
+class AuthController:  # noqa: B903
+    """Controller class for authentication operations."""
 
-    return auth
+    username: str
+    password: str
+    system: str
+    driver: WebDriver
+    wait: WebDriverWait
+    elements: type_elements
+    prt: PrintMessage
+    subclasses = {}
+
+    def __init__(
+        self,
+        username: str,
+        password: str,
+        system: str,
+        driver: WebDriver,
+        wait: WebDriverWait,
+        elements: type_elements,
+        prt: PrintMessage,
+    ) -> None:
+        """Inicializador do AuthController."""
+        self.username = username
+        self.password = password
+        self.system = system
+        self.driver = driver
+        self.wait = wait
+        self.elements = elements
+        self.prt = prt
+
+    @classmethod
+    def construct(cls, system: str, *args, **kwargs) -> Self:  # noqa: ANN002, ANN003
+        """Método construtor para instanciar a classe."""
+        return cls.subclasses.get(system.lower())(*args, **kwargs)
+
+    def __init_subclass__(cls) -> None:  # noqa: D105
+        if not hasattr(cls, "auth"):
+            raise NotImplementedError(
+                f"Subclasses of {cls.__name__} must implement the 'auth' method."
+            )
+
+        cls.subclasses[cls.__name__.lower()] = cls
+
+    def auth(self) -> bool:  # noqa: D102
+        raise NotImplementedError("Method 'auth' must be implemented in subclasses.")
+
+    def accept_cert(self, accepted_dir: str) -> None:  # noqa: D102
+        raise NotImplementedError(
+            "Method 'accept_cert' must be implemented in subclasses."
+        )
