+++ crawjud/addons/auth/elaw.py
@@ -6,10 +6,10 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from crawjud.addons.auth.controller import AuthController
+from crawjud.addons.auth import AuthController
 
 
-class ElawAuth(AuthController):
+class ElawAuth(AuthController, name="elaw"):
     """Classe de autenticação Elaw."""
 
     def auth(self) -> bool:
