+++ crawjud/addons/search/projudi.py
@@ -11,14 +11,14 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from crawjud.addons.search.controller import SearchController
+from crawjud.addons.search import SearchController
 from crawjud.exceptions.bot import ExecutionError
 
 if TYPE_CHECKING:
     from selenium.webdriver.remote.webelement import WebElement
 
 
-class ProjudiSearch(SearchController):
+class ProjudiSearch(SearchController, name="projudi"):
     """Classe de pesquisa PROJUDI."""
 
     def search(self, bot_data: dict[str, str]) -> bool:
